Quickbook: Change the anchor markup for headers.
Use a phrase with an id inside the header.

Using an anchor before the heading was causing problems when the heading
wasn't inside a section.

We can't use an anchor inside the header when it links to itself, as it
messes up getting the title for the link. Using an empty phrase means
that it will just get the empty title for the phrase. Thanks to John
Maddock for the idea.

This does generate slightly odd markup. We could possibly use the old
style markup whenever section depth > 0, since we would know that we're
inside a section, but that seems like an unnecessary complication.

[SVN r74346]